requirements.txt
# Core Python libraries for data science / ML
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2

# Web + app framework
streamlit==1.31.0
validators==0.22.0

# Security / parsing utilities
python-whois==0.8.0
pefile==2023.2.7

# Image handling
Pillow==10.2.0
piexif==1.1.3

# Database
sqlite3-binary==0.0.1  # optional, since Python stdlib already includes sqlite3

# Misc utilities
requests==2.31.0
